requirement.txt
selenium
webdriver-manager
beautifulsoup4
pandas
lxml
requests
numpy
ipykernel
mysql-connector-python
python-dotenv
kaggle
fastapi
uvicorn
jinja2
python-multipart
passlib[bcrypt]==1.7.4 
bcrypt==3.2.2
python-jose[cryptography]
pydantic-settings
email-validator
